+++ src/pyprinttags.py
@@ -15,40 +15,29 @@ of certain tags they know.
 from __future__ import unicode_literals, print_function
 
 import argparse
-import os.path
 import sys
 import taglib
 
+def ensure_unicode(filename: str):
+    return filename.decode(sys.getfilesystemencoding()) if isinstance(filename, bytes) else filename
 
 def script():
-    """Run the command-line script."""
+    """Print tags of given files"""
     parser = argparse.ArgumentParser(description="Print all textual tags of one or more audio files.")
-    parser.add_argument("-b", "--batch", help="disable user interaction", action="store_true")
     parser.add_argument("file", nargs="+", help="file(s) to print tags of")
     args = parser.parse_args()
-    for filename in args.file:
-        if isinstance(filename, bytes):
-            filename = filename.decode(sys.getfilesystemencoding())
-        line = "TAGS OF '{0}'".format(os.path.basename(filename))
-        print("*" * len(line))
-        print(line)
-        print("*" * len(line))
-        audioFile = taglib.File(filename)
-        tags = audioFile.tags
+    for i, filename in enumerate(args.file):
+        filename = ensure_unicode(filename)
+        print('{0}:'.format(filename))
+        file = taglib.File(filename)
+        tags = file.tags
         if len(tags) > 0:
-            maxKeyLen = max(len(key) for key in tags.keys())
+            max_key_len = max(len(key) for key in tags.keys())
             for key, values in tags.items():
                 for value in values:
-                    print(('{0:' + str(maxKeyLen) + '} = {1}').format(key, value))
-        if len(audioFile.unsupported) > 0:
-            print('Unsupported tag elements: ' + "; ".join(audioFile.unsupported))
-            if sys.version_info[0] == 2:
-                inputFunction = raw_input
-            else:
-                inputFunction = input
-            if not args.batch and inputFunction("remove unsupported properties? [yN] ").lower() in ["y", "yes"]:
-                audioFile.removeUnsupportedProperties(audioFile.unsupported)
-                audioFile.save()
+                    print(('  {0:' + str(max_key_len) + '} = {1}').format(key, value))
+        if i < len(args.file) - 1:
+            print()
 
 if __name__ == '__main__':
     script()