+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Changelog
 =========
+unreleased
+----------
+- fix #88: rename pyprinttags3 --> pyprinttags, ignore unsupported tags in the script
 
 pytaglib 1.4.6 (2020/02/26)
 -----------------